# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Modular House API (MVP)
  version: 0.1.0
  description: REST API for public content, enquiry submissions, and lightweight admin.
servers:
  - url: https://api.staging.modular.house
  - url: http://localhost:8080
paths:
...
  /admin/settings/password:
    put:
      tags: [settings]
      summary: Change own password (settings page)
      description: >
        Requires the correct current password plus the new password twice (must match + meet policy).
        On success, revokes the account's other sessions; the acting session stays valid.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed; other sessions revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NeutralAck' }
        '400':
          description: Policy violation, mismatch, or wrong current password.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: No valid session.
        '403':
          description: super_admin account is read-only in the panel.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /admin/settings/photo:
    get:
      tags: [settings]
      summary: Retrieve the current profile photo bytes
      description: >
        Returns the authenticated user's profile photo as an image. 404 when none is set, so the
        client renders the initials fallback (G4/G6).
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Profile photo image.
          content:
            image/png:
              schema: { type: string, format: binary }
            image/jpeg:
              schema: { type: string, format: binary }
            image/webp:
              schema: { type: string, format: binary }
        '401':
          description: No valid session.
        '404':
          description: No profile photo set (client renders initials fallback).
    put:
      tags: [settings]
      summary: Upload or replace the profile photo
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
              required: [photo]
      responses:
        '200':
          description: Photo saved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '400':
          description: Unsupported type (not png/jpeg/webp) or over 5 MB.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: No valid session.
        '403':
          description: super_admin account is read-only in the panel.
    delete:
      tags: [settings]
      summary: Remove the profile photo (fall back to initials)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Photo removed; default avatar (initials) used.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401':
          description: No valid session.
        '403':
          description: super_admin account is read-only in the panel.
  /admin/settings/preferences:
    get:
      tags: [settings]
      summary: Read the persisted theme mode and sidebar collapse state
      description: >
        Returns the server-stored preferences for authoritative cross-device load (the cookie/
        localStorage mirror is only the pre-paint cache).
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current preferences.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Preferences' }
        '401':
          description: No valid session.
    put:
      tags: [settings]
      summary: Persist theme mode and sidebar collapse state
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferencesRequest'
      responses:
        '200':
          description: Preferences saved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Preferences' }
        '400':
          description: Invalid themeMode.
        '401':
          description: No valid session.

  # ─── Admin Content (feature-006, re-gated) ────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword, confirmPassword]
      properties:
        currentPassword: { type: string, minLength: 1, maxLength: 128 }
        newPassword: { type: string, minLength: 12, maxLength: 128 }
        confirmPassword: { type: string, minLength: 12, maxLength: 128 }
    PreferencesRequest:
      type: object
      properties:
        themeMode: { type: string, enum: [light, dark, system] }
        sidebarCollapsed: { type: boolean }
    Preferences:
      type: object
      properties:
        themeMode: { type: string, enum: [light, dark, system] }
        sidebarCollapsed: { type: boolean }
...
    Me:
      type: object
      required: [id, email, role, permissions]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string, nullable: true }
        role: { type: string, example: admin }
        permissions:
          type: array
          items: { type: string, example: 'pages:view' }
        hasProfilePhoto: { type: boolean }
        isSuperAdmin: { type: boolean }
        preferences: { $ref: '#/components/schemas/Preferences' }
    NeutralAck:
      type: object
      required: [message]
      properties:
        message: { type: string }
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }